routes post get et put fonctionnelles + resize image ok

+++ backend/controllers/book.js
@@ -1,30 +1,81 @@
 const Book = require("../models/Book");
+const average = require("../fonctions/average");
+const fs = require("fs");
 
 exports.createBook = (req, res, next) => {
-  delete req.body.id;
+  const bookObject = JSON.parse(req.body.book);
+  delete bookObject._id;
+  delete bookObject._userId;
   const book = new Book({
-    ...req.body,
+    ...bookObject,
+    userId: req.auth.userId,
+    imageUrl: `${req.protocol}://${req.get("host")}/images/resized_${
+      req.file.filename
+    }`,
+    averageRating: bookObject.ratings[0].grade,
   });
   book
     .save()
-    .then(() => res.status(201).json({ message: "Livre enregistré !" }))
-    .catch((error) => res.status(400).json({ error }));
+    .then(() => {
+      res.status(201).json({ message: "Livre enregistré !" });
+    })
+    .catch((error) => {
+      res.status(400).json({ error });
+    });
 };
 
 exports.modifyBook = (req, res, next) => {
-  Book.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-    .then(() => res.status(200).json({ message: "Livre modifié !" }))
-    .catch((error) => res.status(400).json({ error }));
+  const bookObject = req.file
+    ? {
+        ...JSON.parse(req.body.book),
+        imageUrl: `${req.protocol}://${req.get("host")}/images/${
+          req.file.filename
+        }`,
+      }
+    : { ...req.body };
+
+  delete bookObject._userId;
+  Book.findOne({ _id: req.params.id })
+    .then((book) => {
+      if (book.userId != req.auth.userId) {
+        res.status(401).json({ message: "Non authorisé !" });
+      } else {
+        Book.updateOne(
+          { _id: req.params.id },
+          { ...bookObject, _id: req.params.id }
+        )
+          .then(() => res.status(200).json({ message: "Livre modifié!" }))
+          .catch((error) => res.status(401).json({ error }));
+      }
+    })
+    .catch((error) => {
+      res.status(400).json({ error });
+    });
 };
 
 exports.deleteBook = (req, res, next) => {
-  Book.deleteOne({ _id: req.params.id })
-    .then(() => res.status(200).json({ message: "Livre supprimé !" }))
-    .catch((error) => res.status(400).json({ error }));
+  Book.findOne({ _id: req.params.id })
+    .then((book) => {
+      if (book.userId != req.auth.userId) {
+        res.status(401).json({ message: "Non authorisé !" });
+      } else {
+        const filename = book.imageUrl.split("/images/")[1];
+        fs.unlink(`images/${filename}`, () => {
+          Book.deleteOne({ _id: req.params.id })
+            .then(() => {
+              res.status(200).json({ message: "Livre supprimé !" });
+            })
+            .catch((error) => res.status(401).json({ error }));
+        });
+      }
+    })
+    .catch((error) => {
+      res.status(500).json({ error });
+    });
 };
 
 exports.getOneBook = (req, res, next) => {
-  Thing.findOne({ _id: req.params.id })
+  Book.findOne({ _id: req.params.id })
     .then((book) => res.status(200).json(book))
     .catch((error) => res.status(404).json({ error }));
 };
@@ -33,4 +84,53 @@ exports.getAllBooks = (req, res, next) => {
   Book.find()
     .then((book) => res.status(200).json(book))
     .catch((error) => res.status(400).json({ error }));
+};
+
+exports.createRating = (req, res, next) => {
+  if (0 <= req.body.rating <= 5) {
+    const ratingObject = { ...req.body, grade: req.body.rating };
+    delete ratingObject._id;
+    Book.findOne({ _id: req.params.id })
+      .then((book) => {
+        const newRatings = book.ratings;
+        const userIdArray = newRatings.map((rating) => rating.userId);
+        if (userIdArray.includes(req.auth.userId)) {
+          res.status(403).json({ message: "Non authorisé !" });
+        } else {
+          newRatings.push(ratingObject);
+          const grades = newRatings.map((rating) => rating.grade);
+          const averageGrades = average.average(grades);
+          book.averageRating = averageGrades;
+          Book.updateOne(
+            { _id: req.params.id },
+            {
+              ratings: newRatings,
+              averageRating: averageGrades,
+              _id: req.params.id,
+            }
+          )
+            .then(() => {
+              res.status(201).json();
+            })
+            .catch((error) => {
+              res.status(400).json({ error });
+            });
+          res.status(200).json(book);
+        }
+      })
+      .catch((error) => {
+        res.status(404).json({ error });
+      });
+  } else {
+    res
+      .status(400)
+      .json({ message: "La note doit être comprise entre 1 et 5" });
+  }
+};
+exports.getBestRating = (req, res, next) => {
+  Book.find()
+    .sort({ averageRating: -1 })
+    .limit(3)
+    .then((books) => res.status(200).json(books))
+    .catch((error) => res.status(404).json({ error }));
 };
